+++ Discoverer/Injected.swift
@@ -6,6 +6,12 @@
 //  Copyright © 2020 LionSoftware.org. All rights reserved.
 //
 
+/**
++ A property wrapper used to retrieve injected service from injector.
++ Example usage:
++ @Injected
++ var serviceAInjection: PService
++ */
 @propertyWrapper
 public class Injected<Service> {
     private let injector: Injector
@@ -13,6 +19,10 @@ public class Injected<Service> {
     public var wrappedValue: Service {
         //swiftlint:disable force_try
         return try! injector.get(Service.self)
+    }
+    
+    public init() {
+        self.injector = Injector.default
     }
     
     public init(injector: Injector) {