+++ Discoverer/Registered.swift
@@ -6,6 +6,12 @@
 //  Copyright © 2020 LionSoftware.org. All rights reserved.
 //
 
+/**
+ A property wrapper used to register injections into the injectors.
+ Example usage:
+ @Registered
+ var serviceAInjection = Injection<PServiceA>.singleton(ServiceA())
+ */
 @propertyWrapper
 public class Registered<Service> {
     private let injector: Injector
@@ -17,10 +23,22 @@ public class Registered<Service> {
         }
     }
     
+    public init(wrappedValue: Injection<Service>) {
+        self.injector = Injector.default
+        self.injected = wrappedValue
+        //swiftlint:disable force_try
+        try! injector.register(as: Service.self, injection: wrappedValue)
+    }
+
+    
     public init(wrappedValue: Injection<Service>, inInjector injector: Injector) {
         self.injector = injector
         self.injected = wrappedValue
         //swiftlint:disable force_try
         try! injector.register(as: Service.self, injection: wrappedValue)
+    }
+    
+    deinit {
+        try! injector.unregister(type: Service.self)
     }
 }